requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.3
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
backports.zstd==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.42.16
botocore==1.42.16
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
executing==2.2.1
fastjsonschema==2.21.2
fiftyone==1.11.0
fiftyone-brain==0.21.4
fiftyone_db==1.4.0
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
glob2==0.7
graphql-core==3.2.7
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
inflate64==1.0.4
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lia-web==0.2.3
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.14.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.1.0
pybcj==1.0.7
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydash==8.0.5
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.1
pyparsing==3.3.1
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rarfile==4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
rtree==1.4.1
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.48.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8.1
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.50.0
strawberry-graphql==0.287.3
sympy==1.14.0
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.12.20
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.6.2
voxel51-eta==0.15.1
wandb==0.23.1
wcwidth==0.2.14
wrapt==2.0.1
wsproto==1.3.2
xmltodict==1.0.2
